Old/WhatNowAI_test/requirements.txt
# Enhanced Requirements for WhatNowAI with Advanced Features
# =========================================================

# Core web framework
Flask==2.3.3
gunicorn==21.2.0

# API integrations
openai==1.3.5
requests==2.31.0

# Web scraping and parsing
beautifulsoup4==4.12.2
lxml==4.9.3
selenium==4.15.2
cloudscraper==1.2.71

# Text-to-speech
edge-tts==6.1.9

# HTTP client and async support
aiohttp==3.9.1
urllib3==2.0.7

# Search and OSINT capabilities
duckduckgo-search==3.9.6
fake-useragent==1.4.0

# Natural Language Processing (Optional but recommended)
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3

# AI and Machine Learning (Optional)
sentence-transformers==2.2.2
transformers==4.35.2
torch>=2.0.0

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0
colorama==0.4.6
dnspython==2.4.2

# Database and Caching (Optional)
redis==5.0.1

# Development and testing
tqdm==4.66.1
